test: split fast unit lane from script integration lane
Real git clone/worktree/push/fetch flows in the pages-state-branch suite cross
process and filesystem boundaries and routinely exceeded Jest's default 5s
timeout under suite load, making the default test run slow and flaky.

- Scope `npm test` to `tests/unit/` as the fast lane and add a dedicated
  `test:integration:scripts` lane that runs in band with a 30s timeout.
- Move the git-subprocess pages-state-branch coverage into
  `tests/integration/scripts/`, leaving only argument-parsing unit coverage
  behind.
- Run both lanes separately in CI and document them in DEVELOPMENT.md.
