+++ package.json
@@ -34,7 +34,9 @@
     "postman:live-provider": "node scripts/run-postman-live.js",
     "postman:post-deploy": "node scripts/run-postman-deploy.js",
     "report:allure": "npm run allure:clean && npm run test:allure && npm run postman:full:allure && npm run allure:generate",
-    "test": "jest",
+    "test": "jest --coverage=false --testPathPattern=tests/unit/",
+    "test:unit": "jest --coverage=false --testPathPattern=tests/unit/",
+    "test:integration:scripts": "jest --coverage=false --runInBand --testPathPattern=tests/integration/scripts/ --testTimeout=30000",
     "test:allure": "ALLURE_RESULTS_DIR=reports/allure-results jest",
     "test:ci": "jest --ci --reporters=default --reporters=jest-junit",
     "prod": "NODE_ENV=production node src/app.js",
