+++ tests/unit/scripts/github/syncPagesStateBranch.test.js
@@ -1,53 +1,6 @@
-const fs = require('node:fs/promises');
-const os = require('node:os');
 const path = require('node:path');
-const { execFileSync } = require('node:child_process');
 
-const {
-  fetchRemoteBranch,
-  hasRemoteBranch,
-  hasStagedChanges,
-  parseArgs,
-  syncPagesStateBranch,
-  runGit: runGitCommand,
-} = require('../../../../scripts/github/sync-pages-state-branch');
-
-function runGit(cwd, args) {
-  return execFileSync('git', args, {
-    cwd,
-    encoding: 'utf8',
-  }).trim();
-}
-
-async function createTempDir(prefix) {
-  return fs.mkdtemp(path.join(os.tmpdir(), prefix));
-}
-
-async function writeFile(targetPath, contents) {
-  await fs.mkdir(path.dirname(targetPath), { recursive: true });
-  await fs.writeFile(targetPath, contents, 'utf8');
-}
-
-async function createRemoteRepo() {
-  const tempDir = await createTempDir('sync-pages-state-branch-');
-  const remoteDir = path.join(tempDir, 'remote.git');
-  const repoDir = path.join(tempDir, 'repo');
-  const siteDir = path.join(tempDir, 'site');
-
-  runGit(tempDir, ['-c', 'init.defaultBranch=main', 'init', '--bare', remoteDir]);
-  runGit(tempDir, ['clone', remoteDir, repoDir]);
-  runGit(repoDir, ['checkout', '-b', 'main']);
-  await writeFile(path.join(repoDir, 'README.md'), '# repo\n');
-  runGit(repoDir, ['-c', 'user.name=Test User', '-c', 'user.email=test@example.com', 'add', 'README.md']);
-  runGit(repoDir, ['-c', 'user.name=Test User', '-c', 'user.email=test@example.com', 'commit', '-m', 'init']);
-  runGit(repoDir, ['push', '--set-upstream', 'origin', 'main']);
-
-  return {
-    repoDir,
-    rootDir: tempDir,
-    siteDir,
-  };
-}
+const { parseArgs } = require('../../../../scripts/github/sync-pages-state-branch');
 
 describe('Unit | Scripts | GitHub | Sync Pages State Branch', () => {
   it('parses supported CLI arguments', () => {
@@ -114,200 +67,5 @@ describe('Unit | Scripts | GitHub | Sync Pages State Branch', () => {
       '--unsupported',
       'nope',
     ])).toThrow('Unsupported argument: --unsupported');
-  });
-
-  it('exposes git helpers for failure-tolerant checks', async () => {
-    const {
-      repoDir,
-      rootDir,
-    } = await createRemoteRepo();
-
-    try {
-      expect(hasRemoteBranch({
-        branch: 'gh-pages',
-        repoDir,
-      })).toBe(false);
-
-      expect(hasStagedChanges(repoDir)).toBe(false);
-      await writeFile(path.join(repoDir, 'notes.txt'), 'draft');
-      runGit(repoDir, ['add', 'notes.txt']);
-      expect(hasStagedChanges(repoDir)).toBe(true);
-
-      expect(() => runGitCommand({
-        args: ['rev-parse', '--verify', 'refs/heads/does-not-exist'],
-        cwd: repoDir,
-      })).toThrow('git rev-parse --verify refs/heads/does-not-exist failed');
-
-      const allowFailureResult = runGitCommand({
-        allowFailure: true,
-        args: ['rev-parse', '--verify', 'refs/heads/does-not-exist'],
-        cwd: repoDir,
-      });
-      expect(allowFailureResult.status).not.toBe(0);
-    } finally {
-      await fs.rm(rootDir, {
-        force: true,
-        recursive: true,
-      });
-    }
-  });
-
-  it('creates and updates the state branch', async () => {
-    const {
-      repoDir,
-      rootDir,
-      siteDir,
-    } = await createRemoteRepo();
-
-    try {
-      await writeFile(path.join(siteDir, 'index.html'), 'first-report');
-      await writeFile(path.join(siteDir, '.nojekyll'), '');
-
-      const firstSync = await syncPagesStateBranch({
-        branch: 'gh-pages',
-        commitMessage: 'docs: sync first report',
-        repoDir,
-        siteDir,
-      });
-
-      expect(firstSync.changed).toBe(true);
-      expect(firstSync.commitSha).toMatch(/^[a-f0-9]{40}$/u);
-      expect(hasRemoteBranch({
-        branch: 'gh-pages',
-        repoDir,
-      })).toBe(true);
-
-      const publishedCloneDir = path.join(rootDir, 'published');
-      runGit(rootDir, ['clone', '--branch', 'gh-pages', path.join(rootDir, 'remote.git'), publishedCloneDir]);
-      await expect(fs.readFile(path.join(publishedCloneDir, 'index.html'), 'utf8')).resolves.toBe('first-report');
-
-      const secondSync = await syncPagesStateBranch({
-        branch: 'gh-pages',
-        commitMessage: 'docs: sync first report',
-        repoDir,
-        siteDir,
-      });
-
-      expect(secondSync).toEqual({
-        branch: 'gh-pages',
-        changed: false,
-        commitSha: '',
-      });
-
-      await writeFile(path.join(siteDir, 'index.html'), 'second-report');
-
-      const thirdSync = await syncPagesStateBranch({
-        branch: 'gh-pages',
-        commitMessage: 'docs: sync second report',
-        repoDir,
-        siteDir,
-      });
-
-      expect(thirdSync.changed).toBe(true);
-
-      runGit(publishedCloneDir, ['pull', '--ff-only']);
-      await expect(fs.readFile(path.join(publishedCloneDir, 'index.html'), 'utf8')).resolves.toBe('second-report');
-    } finally {
-      await fs.rm(rootDir, {
-        force: true,
-        recursive: true,
-      });
-    }
-  });
-
-  it('fetches and updates an existing remote state branch that is not present locally yet', async () => {
-    const {
-      repoDir,
-      rootDir,
-      siteDir,
-    } = await createRemoteRepo();
-
-    try {
-      await writeFile(path.join(siteDir, 'index.html'), 'first-report');
-      await writeFile(path.join(siteDir, '.nojekyll'), '');
-
-      await syncPagesStateBranch({
-        branch: 'gh-pages',
-        commitMessage: 'docs: sync first report',
-        repoDir,
-        siteDir,
-      });
-
-      const freshCloneDir = path.join(rootDir, 'fresh-clone');
-      runGit(rootDir, ['clone', path.join(rootDir, 'remote.git'), freshCloneDir]);
-      runGit(freshCloneDir, ['update-ref', '-d', 'refs/remotes/origin/gh-pages']);
-
-      expect(hasRemoteBranch({
-        branch: 'gh-pages',
-        repoDir: freshCloneDir,
-      })).toBe(false);
-
-      fetchRemoteBranch({
-        branch: 'gh-pages',
-        repoDir: freshCloneDir,
-      });
-
-      expect(hasRemoteBranch({
-        branch: 'gh-pages',
-        repoDir: freshCloneDir,
-      })).toBe(true);
-
-      await writeFile(path.join(siteDir, 'index.html'), 'second-report');
-
-      const syncResult = await syncPagesStateBranch({
-        branch: 'gh-pages',
-        commitMessage: 'docs: sync second report',
-        repoDir: freshCloneDir,
-        siteDir,
-      });
-
-      expect(syncResult.changed).toBe(true);
-
-      const publishedCloneDir = path.join(rootDir, 'published-state');
-      runGit(rootDir, ['clone', '--branch', 'gh-pages', path.join(rootDir, 'remote.git'), publishedCloneDir]);
-      await expect(fs.readFile(path.join(publishedCloneDir, 'index.html'), 'utf8')).resolves.toBe('second-report');
-    } finally {
-      await fs.rm(rootDir, {
-        force: true,
-        recursive: true,
-      });
-    }
-  });
-
-  it('writes GitHub Actions outputs when executed as a CLI', async () => {
-    const {
-      repoDir,
-      rootDir,
-      siteDir,
-    } = await createRemoteRepo();
-
-    try {
-      const outputFile = path.join(rootDir, 'github-output.txt');
-      await writeFile(path.join(siteDir, 'index.html'), 'cli-report');
-
-      execFileSync('node', [
-        path.join(process.cwd(), 'scripts/github/sync-pages-state-branch.js'),
-        '--repo-dir',
-        repoDir,
-        '--site-dir',
-        siteDir,
-        '--commit-message',
-        'docs: sync cli report',
-        '--output-file',
-        outputFile,
-      ], {
-        cwd: process.cwd(),
-        encoding: 'utf8',
-      });
-
-      await expect(fs.readFile(outputFile, 'utf8')).resolves.toMatch(
-        /^branch=gh-pages\nchanged=true\ncommit_sha=[a-f0-9]{40}\n$/u,
-      );
-    } finally {
-      await fs.rm(rootDir, {
-        force: true,
-        recursive: true,
-      });
-    }
   });
 });
